Criação do User Repository | classe aulixiar de configuração | injeção de dependência | Camada de serviço | registro de componentes

+++ src/main/java/com/educandoweb/curso/entities/User.java
@@ -12,7 +12,7 @@ public class User {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private long id;
+	private Integer id;
 	
 	private String name;
 	
@@ -23,7 +23,7 @@ public class User {
 	public User () {
 	}
 
-	public User(long id, String name, String phone, String password) {
+	public User(Integer id, String name, String phone, String password) {
 		super();
 		this.id = id;
 		this.name = name;
@@ -35,7 +35,7 @@ public class User {
 		return id;
 	}
 
-	public void setId(long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
